+++ swak/jsonobject/fields/__init__.py
@@ -5,11 +5,13 @@ from .flexidate import FlexiDate
 from .flexitime import FlexiTime
 from .custom import CustomField
 from .resolve import resolve
+from .normalizers import Lower
 
 __all__ = [
     'Maybe',
     'FlexiDate',
     'FlexiTime',
     'CustomField',
-    'resolve'
+    'Lower',
+    'resolve',
 ]